Move UserNav into auth folder
